Support mutually exclusive properties in schema
There are several places where resources have mutually exclusive
properties and every such resource has to implement custom validate
logic (one example is any pair of supported
and deprecated properties).

This patch adds a 'conflicts' element to property schema,
that accepts a list of other properties the given one is
mutually exclusive with, and checks that no such propeties
are used together when doing properties validation.
The same logic is added for parameters in order to support
nested stacks and provider resources.

As a matter of support for developers an exception is also raised
during validation when a property defined as conflicting
is not actually defined in the schema or properties having conflicts
defined are not *mutually* exclusive (i.e. conflicting properties
must be present in each other 'conflicts' lists).

Change-Id: I5e9458e38a2b680892975ad9cd2b2c0c1df9dcb6
Implements: bp mutually-exclusive-properties